third_party/WebKit/LayoutTests/editing/execCommand/switch-multiple-list-items-crash.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
  '<div contenteditable>',
    '<a href="z">^<br></a>',
    '<ol>',
      '<li>',
        '<a href="Z"><img>z</a>',
        '<hr>',
      '|</li>',
    '</ol>',
  '</div>'
  ],
  'insertunorderedlist',
  [
  '<div contenteditable>',
    '<ul>',
      '<li>',
        '<font color="#0000ee"><u>^<br></u></font>',
      '</li>',
      '<li>',
        '<a href="Z"><img>z</a><hr>|',
      '</li>',
    '</ul>',
  '</div>'
  ]),
  "This InsertOrderedList switches the type of a list with multiple list items");
</script>
